#44 part 1.1
Part 1.1 tuple from questions 2 to 6 was performed. In question 3, it was solved by considering as a result a new tuple or modified tuple, because it would not be the same tuple directly.
In the question, the step by step of how the result can be obtained is the total sum of the elements of the tuple.