#115
In this assignment, Victor Raico and Renato Rivera elaborate the two functions for extracting tables and select the options from the second dropdown menu. In order to get the data from all the tables in the JNE Web Page, they make a function that iterate all the process to extract the information from the tables in all the presidential elections. Ignacio Huaroto make the chunk to concatenate the dataframes into a main one. Finally he divided the process into chunks and traslated to english.