#1079 AndresVillacorta
I was in charge of handling the raster files, which are essentially satellite images. My job was to connect these raster files with the geographic data (the district boundaries) using a method called zonal statistics. This allowed me to see how the information in the raster—like construction density—was distributed across the different districts. Once I got those stats, I added them to a table (GeoDataFrame), which makes it easier to analyze later.